[SC-5088] Fix resource leaks in async and make ignoreCorruptFiles work for async.
## What changes were proposed in this pull request?
`FileScanRDD` currently leaks file handles in async IO mode. In async mode `FileScanRDD`s readers are created in a different thread. This is a problem when the creation thread has no `TaskContext` or the wrong `TaskContext` associated with it, as a result completion handlers (which handle resource clean up) would not be called for these readers.

This PR fixes this by explicitly setting the `TaskContext` in the reader creating thread.

This PR also adds support for `ignoreCorruptFiles` in async mode. I would have prefered to keep this separate, but was needed to pass tests.

## How was this patch tested?
Existing tests pass (again).

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #120 from hvanhovell/SC-5088.