Re-enable streaming tests.
## What changes were proposed in this pull request?
Currently streaming tests fail often because multiple SparkContexts are created in a single VM. This is both a problem in DB Spark as Apache Spark. This PR fixes by making sure the previous SparkContext is always destroyed before a new one is created, and preferably by reusing the existing SparkContext.

I went kind of overboard on the refactor.

## How was this patch tested?
It are only tests.

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #133 from hvanhovell/fix_streaming_tests.