[SC-4878] Hacky fix for log viewer's "load more" / "load previous" links
## What changes were proposed in this pull request?

This patch attempts to fix the broken "load more" / "load previous" links in the Spark web UI's log viewer page. In a nutshell, the problem is that some of the links and URLs that power this page are constructed in Javascript and thus are not visible to SparkUIProxyServlet's URL-rewriting code, causing them to break.

A partial fix, implemented by me in 600abffb19a2cb5feccae9d78666f958e30b7da2, is to change the URLs to use the relative path `../log/*` instead of `/log`. The reason that this work is that the log viewer Javascript is only served from the master / worker web UIs at `uiroot/logPage`, so the `../` takes us back to the UI root and the `/log` re-descends back into the endpoint for loading more logs. By using relative paths here we ensure that the browser issues the right requests from inside of the IFrame.

That earlier patch of mine fixed the load more link when navigating to the log viewer page from the driver UI's executor page links, but it turns out that the load more links are still broken when navigating to that same log viewer page from the worker web UI or master web UI. It turns out that this is caused by a subtle difference in behavior due to _how_ those pages are linked to:

- The driver UI generates links with a trailing slash after the `logPage` part, e.g.  `logPage/?appId=%s&executorId=%s&logType=stdout`.
- The master and worker UIs, however, omit the trailing slash and generate URLs like `logPage?appId=%s&executorId=%s&logType=stdout`; these URLs don't work with the `../` trick because they cause the generated URL / request to go up one level too high and miss the SparkContextID parameter, resulting in a broken URL which returns an error response and causes a Javascript error.

The new fix, implemented here, is still somewhat hacky: I just updated the URL generation code to always append that trailing slash. This risks breakage in case the open-source UI changes significantly in the future, but we're already vulnerable to such breakage and the more comprehensive fix will involve either extensive upstream changes to how we generate URLs in the UI (e.g. always fetching the base URL from an attribute of a HTML element on the page and using that in the JS). (Also, before anyone suggests this, the HTML5 `<base>` tag is not a good idea here since it risks serious breakage of anchor links and third-party JS).

## How was this patch tested?

Manually on Dogfood via the following steps:

- From the cluster page, navigate to the driver UI:
  - Click executors tab, then for the first executor:
    - Click 'stderr', verify that log viewer works. Whether this works or not can be verified by clicking on "load more": it should either display more output or should put up blue informational text saying that no more output is available.
    - Go back, click 'stdout', test that page.
- From the cluster page, click on Spark master UI link.
  - Click on application details page for the only running application.
    - Test stdout and stderr pages for the first executor.
- From the cluster page, click on Spark worker UI link:
  - Test sdtout and stderr pages for the first executor.

Author: Josh Rosen <joshrosendatabricks.com>

(cherry picked from commit dfd994f7121517c50d5926ccd0715843b0d548b7)
Signed-off-by: Herman van Hovell <hvanhovelldatabricks.com>

## Conflicts:
	core/src/main/scala/org/apache/spark/deploy/master/ui/ApplicationPage.scala

Author: Josh Rosen <joshrosen@databricks.com>

Closes #122 from hvanhovell/SC-4878.