[SC-3623][BRANCH-2.1] Dynamic partition pruning
## What changes were proposed in this pull request?
This PR ports https://github.com/databricks/spark/pull/31 over to DB Spark branch-2.1. This adds dynamic partition pruning (see the original PR for more details on the feature).

This was non-trivial to port because the read path has changed significantly in Spark 2.1. We only support partition pruning for `HadoopFsRelation`. This relation is exclusively read using the `FileSourceScanExec` and so I have only implemented dynamic partition pruning for this scan operator (in 2.0 we support dynamic partition pruning for both `RowScanExec` and `FileSourceScanExec`).

## How was this patch tested?
Added a test to `SQLQuerySuite`.

Author: Herman van Hovell <hvanhovell@databricks.com>
Author: Davies Liu <davies.liu@gmail.com>

Closes #131 from hvanhovell/dynamic_partition_pruning.