[SC-5452] Do not assume that an ExternalCatalog is a HiveExternalCatalog
## What changes were proposed in this pull request?
The Spark extensions suite added a wrapper for the `ExternalCatalog`. This wrapper is injected into the `SharedState` when the spark extensions contain catalog hooks, and it sits between the `SharedState` and the actual `ExternalCatalog`. This - unfortunately - breaks `InsertIntoHiveTable` and a number tests, that assumes that the `ExternalCatalog` in the `SharedState` is always a `HiveExternalCatalog` in order to extract the `HiveClient`.

This PR fixes this problem by adding a utility method that properly extracts the `HiveClient` from the `SharedState`, and it updates all uses of the `spark.sharedState.externalCatalog.asInstanceOf[HiveExternalCatalog].client` with the new utility method.

## How was this patch tested?
Existing tests.

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #153 from hvanhovell/SC-5452.