[SC-5483] Fix CreateTempTable/ShowPermissions/ShowDatabases in CheckPermissions
## What changes were proposed in this pull request?
The CheckPermissions rule currently does not handle  the `ShowDatabasesCommand` and the `ShowPermissionsCommand` nodes, as a result the rule will throw a SecurityException when one of these commands is used. This PR adds these commands to the whitelist.

We also change the required permission for `CreateTempTableUsing` and the 'run-SQL-on-file' functionality, both now require a Select on an Anonymous File (a blanket permission). The reason for this is that both code paths allow a user to interact directly with the filesystem, and gives the user a way to circumvent ACLs defined on a table (by querying the files backing the table directly).

## How was this patch tested?
Added tests to `CheckPermissionRuleSuite`.

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #156 from hvanhovell/SC-5483.