[SC-5606] Inline spark-avro sources into databricks/spark
This patch ports `spark-avro` as of https://github.com/databricks/spark-avro/commit/b01a034b0e5d785609275aa97d9c5e3719613194 and updates it to run with Spark 2.10 by including the fixes from https://github.com/databricks/spark-avro/pull/206.

Via a transitive dependency on `spark-core` and `spark-sql`, this winds up not needing to add new dependencies on Avro to the poms.

I've updated the license headers to use the header that we use for Spark-edge features.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #170 from JoshRosen/add-spark-avro.