[SPARK-18917][SC-5624] Remove schema check in appending data
In append mode, we check whether the schema of the write is compatible with the schema of the existing data. It can be a significant performance issue in cloud environment to find the existing schema for files. This patch removes the check.

Note that for catalog tables, we always do the check, as discussed in #16339 (comment)

backport https://github.com/apache/spark/pull/16622 to our internal 2.1 branch.

Author: Reynold Xin <rxin@databricks.com>

Closes #178 from cloud-fan/backport.