[SC-5603] Make CTAS work with ACL enable DB Spark
## What changes were proposed in this pull request?
ACL enabled Spark currently throws a table not found exception when we issue a CTAS command. This bug is caused by a number of factors:

1. CTAS is composed of two stages: creating the table and inserting data into the table.
2. The order of these commands has been inverted to make sure we never expose a partially written table to an end user.
3. We check ACLs for the CTAS command, and also for its child commands. ACL checking fails because we cannot find the table when executing the insert (it has not been created yet).

This PR fixes this by only ACL checking the CTAS command, and not its children. This fix is also applied to all other `RunnableCommand`s. The only exception is the `ExplainCommand`, in this case we **want** to execute `CheckAnalysis` on its child.

This has been implemented by adding a trusted mode to `CheckPermissions`. We do not check the permissions when we are in trusted mode. The value of trusted mode is thread local, and can by set by calling the `CheckPermissions.trusted` method. Note that a creative developer can abuse this by either using Java (the trusted method is package-local to `com.databricks`) or some light reflection. The other part of the solution is that we add a planning rule to `QueryExecution`, which wraps a `RunnableCommand` with a `TrustedRunnableCommand `. Note that the downside to this is that we are introducing a change in an apache package which might create merge conflicts.

Finally I have added an sbt task `sparkShellACL` which starts the Spark Shell with ACLs enabled. This is easier for testing. Remember to set `sc.setLocalProperty("spark.databricks.token", "super")` when you start the shell (if you don't every command will fail). You can change users by modifying the `spark.databricks.token` local spark properties.

## How was this patch tested?
Existing tests & Added an integration test to the ACL end to end tests.

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #184 from hvanhovell/SC-5603.