[SC-5584][SC-5585][SC-5583] Auto-vacuum, vacuum <table>, and vacuum ACL support
## What changes were proposed in this pull request?

This extends the VACUUM command to allow referencing both a table and path. When a table is vacuumed, we gather partition stats and generate a minimal set of paths that cover all the partitions of the table.

We require ANY FILE permission to issue VACUUM on a path, and MODIFY permission on the table to issue VACUUM on a table.

In addition, we auto-vacuum updated paths after job commits succeed. This avoids indefinite buildup of files when a directory is repeatedly overwritten, even without manual intervention.

### Config flags:
com.databricks.sql.autoVacuumOnCommit -- whether we should auto vacuum (default true)
com.databricks.sql.defaultVacuumRetentionHours -- number of hours to retain for auto vacuum (default 48.0)

## How was this patch tested?

Unit tests, also verified manually on dogfood.

Author: Eric Liang <ekl@databricks.com>
Author: Eric Liang <ekhliang@gmail.com>

Closes #168 from ericl/auto-vacuum.