[SC-5832] Move all edge configurations into a single file
## What changes were proposed in this pull request?
We currently just interleave edge specific configurations with normal Spark configurations, making it very difficult to track the list of Databricks-specific configs we have and increase the chance of merge conflicts with upstream apache.

This patch moves all edge configurations into a single file in com.databricks.sql package.

## How was this patch tested?
This should work as long as it compiles.

Author: Reynold Xin <rxin@databricks.com>

Closes #215 from rxin/SC-5832.