[SC-5704][REDSHIFT] Catalyst Expression to Redshift SQL converter
## What changes were proposed in this pull request?

This patch implements a Catalyst Expression to Redshift SQL converter.

This is one largely independent piece of the [larger PR](https://github.com/databricks/spark/pull/209) for implementing advanced query pushdown to Redshift.

It basically consists of a collection of "supported" expressions that are organized in a number of partial functions and a main `convert()` method that takes an `Expression` and, if there is a match, turns it into a piece of SQL text that Redshift can understand.

The code is currently unused, except for the Unit tests below.

## How was this patch tested?

Added `ExpressionToSqlSuite.scala` Unit tests.

Author: Adrian Ionescu <adrian@databricks.com>

Closes #221 from adrian-ionescu/redshift-expr-conv.