[SC-5556] Automatically kill queries that are producing too many output rows
# What changes were proposed in this pull request?

This change implements DatabricksTaskDebugListener, which continuously monitors the metric updates send gradually by tasks in execution. In particular, the number of records read and produced is inspected.

The number of records read is computed as the sum of "internal.metrics.input.recordsRead" and "internal.metrics.shuffle.read.recordsRead" metrics.

The number of records produced is the maximum value of "number of output rows" metric. This way we can detect, e.g., cardprod immediately followed by a filter.

The tasks for which the ratio between these two values exceeds configured limit are terminated. The mechanism is off by default, can be turned on by modifying parameter spark.databricks.debug.taskKiller.outputRatioThreshold.

# How was this patch tested?

Contains unit tests.

Author: Ala Luszczak <ala@databricks.com>

Closes #201 from ala/sc-5556.