[SPARK-19650] Commands should not trigger a Spark job
This is a backport of https://github.com/apache/spark/commit/8f0511ed49a353fb0745f320a84063ced5cc1857.

## What changes were proposed in this pull request?
Spark executes SQL commands eagerly. It does this by creating an RDD which contains the command's results. The downside to this is that any action on this RDD triggers a Spark job which is expensive and is unnecessary.

This PR fixes this by avoiding the materialization of an `RDD` for `Command`s; it just materializes the result and puts them in a `LocalRelation`.

## How was this patch tested?
Added a regression test to `SQLQuerySuite`.

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #253 from hvanhovell/SPARK-19650.