[SC-6061] Fix nullability in BNL
## What changes were proposed in this pull request?
`BroadcastNestedLoopJoin` did not check if a condition was null. This causes some surprising bugs when it is used with an `Or` expressions in a `LEFT ANTI JOIN`. The PR addresses this issue.

## How was this patch tested?
It now passes tests :)....

Author: Herman van Hovell <hvanhovell@databricks.com>

Closes #257 from hvanhovell/SC-6061.

(cherry picked from commit 3bd6fd41e00e5890fa3bcd78e4b06fbd8a3bb071)
Signed-off-by: Yin Huai <yhuai@databricks.com>