[SC-6069] FileSourceStrategySuite hangs during directory commit vacuum
## What changes were proposed in this pull request?

The root cause was pool-induced deadlock. Both the VACUUM and resolveCommitState() operation were using a thread-pool, so when a parallel VACUUM attempted to call resolveCommitState(), you could hang indefinitely if there were not enough pool threads available.

The fix is to have the different operations use separate thread-pools.

## How was this patch tested?

Fixes an existing test.

Author: Eric Liang <ekl@databricks.com>

Closes #260 from ericl/sc-6069.